DIPS module finished and cleaned up
